// frontend/src/lib/api/prisma/prisma.interfaces.d.ts
export interface UploadReadyEvent {
    upload_id: string;
    run_id: string;
    node_id: string;
    file_size: number;
    experiment_ids: string[];
    collection_id: string;
    organization_id: string;
    orbit_id: string;
    manifest: Record<string, any>;
    file_index: Record<string, [number, number]>;
}